requirements.txt
numpy
mlflow
pandas
scikit-learn
pandas
urllib3

azureml-pipeline